+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Xiangqi support: variant and piece definitions, flying generals rule.
 - Fixed a bug where the SAN format for pawn captures might be wrong (thanks @malaschitz).
 - Fixed Chess960 castling moves not being generated for kings on g1 (thanks @malaschitz).
+- Fixed Crazyhouse bugs: pawns being droppable on the first rank, and promoted pieces not being captured as pawns (thanks @malaschitz).
 
 ### 0.6.0
 - Support for hopper pieces, such as the Grasshopper and Xiangqi Cannon, and Betza modifiers 'p' and 'g'.